requirements.txt
python-dotenv
pyrebase4
flet==0.80.5 
flet-cli==0.80.5 
flet-desktop==0.80.5 
flet-geolocator==0.80.5 
flet-map==0.80.5 
flet-web==0.80.5